% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alr_results.R
\name{alr_results}
\alias{alr_results}
\title{Visualize Results of Alternating Logistic Regressions.}
\usage{
alr_results(model_fit, title, included_models = "both")
}
\arguments{
\item{model_fit}{an object returned by `ordgee`}

\item{title}{Text for the title portion of your results}

\item{included_models}{which model you want to show in results. Defaults to both mean and correlation. For mean enter "mean", correlation: "correlation".}
}
\description{
Takes output from the `ordgee` function in the package `geepack` and returns a table with visual results along with numerical values imposed on the graph
}
\examples{
alr_results(my_fit, "Results from fitting alternating logistic regression", included_models = "both")
}
